phase3-django-api/pyproject.toml
[project]
name = "phase3-django-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.2",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "psycopg[binary]>=3.3.3",
    "python-decouple>=3.8",
]
